+++ src/Printed/Bundle/Queue/Service/QueueTaskDispatcher.php
@@ -142,13 +142,38 @@ class QueueTaskDispatcher
      * It's also the only safe way to dispatch queue tasks from some of the doctrine listeners
      * (e.g. postUpdate) in your app.
      *
-     * @param AbstractQueuePayload $payload
+     * If you can't create an instance of queue payload before the Doctrine flush (e.g. because you need some ids to pass
+     * in the payload), then pass a "payload creator function", which is invoked after final Doctrine flush but before
+     * the queue task dispatch.
      *
+     * @param AbstractQueuePayload|callable $payloadOrPayloadCreatorFn This must not be a callable in the array form.
+     *      The callable must return an instance of AbstractQueuePayload and require no arguments
      * @return ScheduledQueueTask
      */
-    public function dispatchAfterNextEntityManagerFlush(AbstractQueuePayload $payload): ScheduledQueueTask
+    public function dispatchAfterNextEntityManagerFlush($payloadOrPayloadCreatorFn): ScheduledQueueTask
     {
-        $existingScheduledQueueTask = $this->payloadsDelayedUntilNextDoctrineFlush[spl_object_hash($payload)]
+        /*
+         * Disallow "array" callables because I can't get object hashes of arrays.
+         */
+        if (is_array($payloadOrPayloadCreatorFn)) {
+            throw new \InvalidArgumentException(sprintf(
+                'Callables in the array form are not supported in `%s`.',
+                __METHOD__
+            ));
+        }
+
+        if (
+            !$payloadOrPayloadCreatorFn instanceof AbstractQueuePayload
+            && !is_callable($payloadOrPayloadCreatorFn)
+        ) {
+            throw new \InvalidArgumentException(sprintf(
+                'Argument to `%s` function must either be an instance of `%s` or a callable',
+                __METHOD__,
+                AbstractQueuePayload::class
+            ));
+        }
+
+        $existingScheduledQueueTask = $this->payloadsDelayedUntilNextDoctrineFlush[spl_object_hash($payloadOrPayloadCreatorFn)]
             ?? null;
 
         /*
@@ -158,10 +183,19 @@ class QueueTaskDispatcher
             return $existingScheduledQueueTask;
         }
 
-        $this->throwIfPayloadInvalid($payload);
+        /*
+         * Validate instantiated payloads now. The "delayed" payloads are validated just before they are dispatched.
+         */
+        if ($payloadOrPayloadCreatorFn instanceof AbstractQueuePayload) {
+            $this->throwIfPayloadInvalid($payloadOrPayloadCreatorFn);
+        }
 
-        $scheduledQueueTask = new ScheduledQueueTask($payload);
-        $this->payloadsDelayedUntilNextDoctrineFlush[spl_object_hash($payload)] = $scheduledQueueTask;
+        $scheduledQueueTask = new ScheduledQueueTask(
+            $payloadOrPayloadCreatorFn instanceof AbstractQueuePayload ? $payloadOrPayloadCreatorFn : null,
+            is_callable($payloadOrPayloadCreatorFn) ? $payloadOrPayloadCreatorFn : null
+        );
+
+        $this->payloadsDelayedUntilNextDoctrineFlush[spl_object_hash($payloadOrPayloadCreatorFn)] = $scheduledQueueTask;
 
         return $scheduledQueueTask;
     }
@@ -183,9 +217,21 @@ class QueueTaskDispatcher
         foreach ($this->payloadsDelayedUntilNextDoctrineFlush as $scheduledQueueTask) {
             /** @var ScheduledQueueTask $scheduledQueueTask */
 
+            $payload = $scheduledQueueTask->getPayload();
+            $isPayloadConstructedLate = false;
+
+            if (!$payload) {
+                $payload = $scheduledQueueTask->constructAndGetPayload();
+
+                /*
+                 * I don't like how the payload validation may fail after doctrine flush, but well..
+                 */
+                $isPayloadConstructedLate = true;
+            }
+
             $queueTask = $this->dispatch(
-                $scheduledQueueTask->getPayload(),
-                ['validatePayload' => false]
+                $payload,
+                ['validatePayload' => $isPayloadConstructedLate]
             );
 
             $scheduledQueueTask->setQueueTask($queueTask);